requirements.txt
treamlit
google-generativeai
python-dotenv
langchain
pandas
faiss-cpu
langchain_google_genai
langchain-community